Extract common app code into standalone package
