+++ packages/fractal/fractal/package.json
@@ -8,6 +8,7 @@
   "author": "Mark Perkins <mark@allmarkedup.com> (http://allmarkedup.com)",
   "license": "MIT",
   "dependencies": {
+    "@frctl/app": "^2.0.0-beta.0",
     "@frctl/utils": "^2.0.0-beta.0",
     "@frctl/config": "^2.0.0-beta.0",
     "@frctl/support": "^2.0.0-beta.0",
