+++ packages/fractal/fractal/src/fractal.test.js
@@ -1,15 +1,11 @@
 /* eslint import/no-dynamic-require: off,  no-unused-expressions: off */
 
-const {writeFileSync, mkdirSync} = require('fs');
-const {tmpdir} = require('os');
 const {join} = require('path');
 const {capitalize} = require('lodash');
 const {File, ComponentCollection, FileCollection, EmittingPromise, Component, Variant, Collection} = require('@frctl/support');
 const {defaultsDeep} = require('@frctl/utils');
 const {Renderer} = require('@frctl/renderer');
-const {Parser} = require('@frctl/parser');
-const Cache = require('node-cache');
-const {FSWatcher} = require('chokidar');
+const App = require('@frctl/app');
 const {expect, sinon} = require('../../../../test/helpers');
 const pkg = require('../package.json');
 const ConfigStore = require('./config/store');
@@ -18,10 +14,7 @@ const Fractal = require('./fractal');
 
 const config = {
   src: join(__dirname, '../../../../test/fixtures/components'),
-  extends: null,
-  commands: [
-    './test/fixtures/add-ons/command.js'
-  ],
+  presets: null,
   adapters: [
     './test/fixtures/add-ons/adapter.js'
   ]
@@ -63,23 +56,16 @@ const components = new ComponentCollection([
 
 const parserOutput = {components, files};
 
-const tmp = join(tmpdir(), Date.now().toString());
-
-mkdirSync(tmp);
-
-function writeFile(name, contents) {
-  writeFileSync(join(tmp, name), contents);
-}
-
 function makeFractal(customConfig) {
   return new Fractal(customConfig || config);
 }
 
 describe('Fractal', function () {
   describe('constructor()', function () {
-    it.skip('accepts configuration data', () => {
+    it('wraps configuration data in a ConfigStore instance', () => {
       const fractal = makeFractal();
       expect(fractal.config.data).to.eql(defaultsDeep(config, defaults));
+      expect(fractal.config).to.be.instanceOf(ConfigStore);
     });
 
     it('throws an error if invalid config data is provided', () => {
@@ -89,79 +75,8 @@ describe('Fractal', function () {
     it('does not throw an error if no config data is provided', () => {
       expect(() => new Fractal()).to.not.throw();
     });
-
-    it('sets the dirty flag to true', () => {
-      const fractal = new Fractal();
-      expect(fractal.dirty).to.equal(true);
-    });
-  });
-
-  describe('.get()', function () {
-    it('retrieves a value from the config data', () => {
-      const fractal = makeFractal();
-      expect(fractal.get('foo')).to.equal(config.foo);
-    });
-    it('accepts a fallback argument which is returned if the property is undefined', () => {
-      const fractal = makeFractal();
-      const fallback = 'whoops!';
-      expect(fractal.get('boop', fallback)).to.equal(fallback);
-    });
-  });
-
-  describe('.set()', function () {
-    it('sets a value in the config data', () => {
-      const fractal = makeFractal();
-      fractal.set('foo.bar', 'baz');
-      expect(fractal.get('foo.bar')).to.equal('baz');
-    });
-    it('sets the dirty flag', () => {
-      const fractal = makeFractal();
-      fractal.dirty = false;
-      fractal.set('foo.bar', 'baz');
-      expect(fractal.dirty).to.equal(true);
-    });
-  });
-
-  describe('.parse()', function () {
-    it('returns an EmittingPromise', function () {
-      const fractal = new Fractal();
-      expect(fractal.parse()).to.be.instanceOf(EmittingPromise);
-    });
-    it('resolves to an object with file and component collections', async function () {
-      const fractal = new Fractal();
-      const {components, files} = await fractal.parse();
-      expect(components).to.be.instanceOf(ComponentCollection);
-      expect(files).to.be.instanceOf(FileCollection);
-    });
-    it('uses the cached result if valid', async function () {
-      const fractal = new Fractal();
-      const spy = sinon.spy(fractal.cache, 'get');
-      const collections = await fractal.parse();
-      expect(spy.called).to.equal(true);
-      expect(spy.returned(collections)).to.equal(false);
-      spy.reset();
-      const cachedCollections = await fractal.parse();
-      expect(spy.returned(cachedCollections)).to.equal(true);
-      spy.restore();
-    });
-    it('rejects if an error is thrown by the parser', async function () {
-      const fractal = new Fractal();
-      const stub = sinon.stub(fractal, 'getParser').callsFake(function () {
-        return {
-          run() {
-            return Promise.reject(new Error('oops!'));
-          }
-        };
-      });
-      let tested = false;
-      try {
-        await fractal.parse();
-      } catch (err) {
-        expect(err).to.be.instanceOf(Error);
-        tested = true;
-        stub.restore();
-      }
-      expect(tested).to.equal(true);
+    it('extends App', () => {
+      expect(new Fractal()).to.be.instanceOf(App);
     });
   });
 
@@ -273,145 +188,52 @@ describe('Fractal', function () {
     });
   });
 
-  describe('.watch()', function () {
-    let fractal;
-    let watcher;
-    beforeEach(function () {
-      fractal = new Fractal({
-        src: [tmp + '/**/*']
-      });
-      watcher = fractal.watch();
-    });
-    afterEach(function () {
-      watcher.close();
-    });
-    it('returns a chokidar instance', function () {
-      expect(watcher).to.be.instanceOf(FSWatcher);
-    });
-    it('returns the same instance if called twice', function () {
-      const sameWatcher = fractal.watch();
-      expect(watcher).to.equal(sameWatcher);
-    });
-    it('sets the chokidar opts correctly', function () {
-      expect(watcher.options.ignoreInitial).to.equal(true);
-      expect(watcher.options.cwd).to.equal(process.cwd());
-    });
-    it('sets the dirty flag when the filesystem changes', function (done) {
-      if (process.env.IS_CI) {
-        this.skip();
-      }
-      watcher.on('all', function () {
-        expect(fractal.dirty).to.equal(true);
-        done();
-      });
-      fractal.dirty = false;
-      watcher.on('ready', () => writeFile('foo.js'));
-    });
-  });
-
-  describe('.unwatch()', function () {
-    let fractal;
-    let watcher;
-    beforeEach(function () {
-      fractal = new Fractal({
-        src: [tmp + '/**/*']
-      });
-      watcher = fractal.watch();
-    });
-    afterEach(function () {
-      watcher.close();
-    });
-    it('closes the chokidar instance', function () {
-      fractal.unwatch();
-      expect(watcher.closed).to.equal(true);
-    });
-    it('removes the instance so the next call to watch() initializes a fresh one', function () {
-      fractal.unwatch();
-      const newWatcher = fractal.watch();
-      expect(watcher).to.not.equal(newWatcher);
-    });
-  });
-
-  for (const addOn of ['plugin', 'adapter', 'transform']) {
+  for (const addOn of ['plugin', 'transform', 'adapter']) {
     const method = `add${capitalize(addOn)}`;
-    describe(`${method}()`, function () {
-      it(`adds a plugin to the ${addOn}s config array`, function () {
-        const fractal = new Fractal({extends: null});
-        const length = addOn === 'transform' ? 1 : 0; // TODO find a nicer way to do this
-        expect(fractal.get(`${addOn}s`)).to.be.an('array').and.have.property('length').which.equals(length);
-        fractal[method](`./test/fixtures/add-ons/${addOn}`);
-        expect(fractal.get(`${addOn}s`).length).equal(length + 1);
+    describe(`.${method}()`, function () {
+      it(`adds a ${addOn} to the ${addOn}s config array`, function () {
+        const app = new Fractal({presets: null});
+        expect(app.get(`${addOn}s`)).to.be.an('array').and.have.property('length').which.equals(0);
+        app[method](`./test/fixtures/add-ons/${addOn}`);
+        expect(app.get(`${addOn}s`).length).equal(1);
       });
-      it(`marks the fractal instance as dirty`, function () {
-        const fractal = new Fractal({extends: null});
-        fractal.dirty = false;
-        fractal[method](`./test/fixtures/add-ons/${addOn}`);
-        expect(fractal.dirty).to.equal(true);
+      it(`marks the app instance as dirty`, function () {
+        const app = new Fractal({presets: null});
+        app.dirty = false;
+        app[method](`./test/fixtures/add-ons/${addOn}`);
+        expect(app.dirty).to.equal(true);
       });
       it(`accepts ${addOn} paths`, function () {
-        const fractal = new Fractal({extends: null});
-        fractal[method](`./test/fixtures/add-ons/${addOn}`);
-        expect(fractal.get(`${addOn}s`)[0]).to.be.an(`object`);
+        const app = new Fractal({presets: null});
+        app[method](`./test/fixtures/add-ons/${addOn}`);
+        expect(app.get(`${addOn}s`)[0]).to.be.an(`object`);
       });
       it(`accepts ${addOn} packages`, function () {
-        const fractal = new Fractal({extends: null});
-        fractal[method](require(`../../../../test/fixtures/add-ons/${addOn}`));
-        expect(fractal.get(`${addOn}s`)[0]).to.be.an(`object`);
+        const app = new Fractal({presets: null});
+        app[method](require(`../../../../test/fixtures/add-ons/${addOn}`));
+        expect(app.get(`${addOn}s`)[0]).to.be.an(`object`);
       });
       it(`accepts ${addOn} objects`, function () {
-        const fractal = new Fractal({extends: null});
-        fractal[method](require(`../../../../test/fixtures/add-ons/${addOn}`)());
-        expect(fractal.get(`${addOn}s`)[0]).to.be.an(`object`);
+        const app = new Fractal({presets: null});
+        app[method](require(`../../../../test/fixtures/add-ons/${addOn}`)());
+        expect(app.get(`${addOn}s`)[0]).to.be.an(`object`);
       });
       it(`accepts ${addOn} config via the add-on array syntax`, function () {
-        const fractal = new Fractal({extends: null});
+        const app = new Fractal({presets: null});
         const spy = sinon.spy(() => ({
           name: `test-${addOn}`
         }));
         const opts = {dothis: true};
-        fractal[method]([spy, opts]);
-        expect(fractal.get(`${addOn}s`)[0]).to.be.an(`object`);
+        app[method]([spy, opts]);
+        expect(app.get(`${addOn}s`)[0]).to.be.an(`object`);
         expect(spy.calledWith(opts)).to.equal(true);
       });
-      it(`returns the Fractal instance`, function () {
-        const fractal = new Fractal({extends: null});
-        expect(fractal[method](`./test/fixtures/add-ons/${addOn}`)).to.equal(fractal);
+      it(`returns the App instance`, function () {
+        const app = new Fractal({presets: null});
+        expect(app[method](`./test/fixtures/add-ons/${addOn}`)).to.equal(app);
       });
     });
   }
-
-  describe('.getParser()', function () {
-    it('returns a new Parser instance', function () {
-      const fractal = new Fractal();
-      const parser = fractal.getParser();
-      expect(parser).to.be.instanceOf(Parser);
-      expect(fractal.getParser()).to.not.equal(parser);
-    });
-    it('initialises the parser with src, plugins and transforms from the config', function () {
-      const fractal = new Fractal({
-        src: ['/foo'],
-        plugins: [
-          {
-            name: 'foo-plugin',
-            collection: 'components',
-            handler: function () {}
-          }
-        ],
-        transforms: [
-          {
-            name: 'tests',
-            transform: function () {
-              return new Collection();
-            }
-          }
-        ]
-      });
-      const parser = fractal.getParser();
-      expect(parser.sources.map(src => src.base)[0]).to.equal('/foo');
-      expect(parser.getTransform('tests')).to.have.property('name').that.equals('tests');
-      expect(parser.getTransform('components').plugins.items.map(plugin => plugin.name)).to.include('foo-plugin');
-    });
-  });
 
   describe('.getRenderer()', function () {
     it('returns a new Renderer instance', function () {
@@ -434,42 +256,6 @@ describe('Fractal', function () {
     it('property describes the Fractal instance', function () {
       const fractal = new Fractal();
       expect(fractal.toString()).to.equal('[object Fractal]');
-    });
-  });
-
-  describe('.dirty', function () {
-    it('gets and sets the dirty property', function () {
-      const fractal = new Fractal();
-      fractal.dirty = false;
-      expect(fractal.dirty).to.equal(false);
-      fractal.dirty = true;
-      expect(fractal.dirty).to.equal(true);
-    });
-    it('clears the cache when set to true', function () {
-      const fractal = new Fractal();
-      const spy = sinon.spy(fractal.cache, 'del');
-      fractal.dirty = true;
-      expect(spy.called).to.equal(true);
-      spy.reset();
-      fractal.dirty = false;
-      expect(spy.called).to.equal(false);
-      spy.restore();
-    });
-  });
-
-  describe('.cache', function () {
-    it('returns the parser cache instance', function () {
-      const fractal = new Fractal();
-      expect(fractal.cache).to.be.instanceof(Cache);
-    });
-  });
-
-  describe('.config', function () {
-    it('returns the config instance', function () {
-      const fractal = new Fractal({
-        foo: 'bar'
-      });
-      expect(fractal.config).to.be.instanceOf(ConfigStore);
     });
   });
