+++ packages/internals/app/README.md
@@ -1,0 +1,15 @@
+# Fractal App
+
+Base class for a filesystem-parsing application.
+
+[![NPM Version](https://img.shields.io/npm/v/@frctl/support.svg?style=flat-square)](https://www.npmjs.com/package/@frctl/app)
+
+## Installation
+
+```bash
+npm i @frctl/app --save
+```
+
+## Requirements
+
+Node >= v7.6 is required.
