+++ packages/internals/app/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@frctl/app",
+  "version": "2.0.0-beta.0",
+  "description": "Fractal application class",
+  "homepage": "http://fractal.build",
+  "repository": "https://github.com/frctl/fractal/tree/v2/packages/internals/app",
+  "bugs": "https://github.com/frctl/fractal/issues",
+  "author": "Mark Perkins <mark@allmarkedup.com> (http://allmarkedup.com)",
+  "license": "MIT",
+  "dependencies": {
+    "@frctl/parser": "2.0.0-beta.0",
+    "@frctl/support": "2.0.0-beta.0",
+    "@frctl/config": "2.0.0-beta.0",
+    "@frctl/utils": "2.0.0-beta.0",
+    "check-types": "^7.2.0",
+    "lodash": "^4.17.4",
+    "debug": "^3.0.0",
+    "chokidar": "^1.7.0",
+    "node-cache": "^4.1.1"
+  }
+}
